+++ .gitignore
@@ -1,3 +1,4 @@
 dist
 build
 node_modules
+Sketch Image Compressor.sketchplugin
