+++ assets/compress-full-jpg.sh
