+++ src/manifest.json
@@ -1,0 +1,19 @@
+{
+  "compatibleVersion": 3.8,
+  "bundleVersion": 1,
+  "identifier": "com.sketchapp.plugins.image-compressor",
+  "commands": [
+    {
+      "name": "Export All Assets",
+      "identifier": "imageCompress",
+      "script": "./plugin.js",
+      "handlers": {
+        "run" : "exportAndCompress",
+        "actions": {
+          "ExportSlices": "compressAutomatically",
+          "Export": "compressAutomatically"
+        }
+      }
+    }
+  ]
+}
